refactor: restructure server models, relocate relay and items modules out of server package, update imports accordingly
