+++ src/routes/world.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter, Depends, HTTPException, Security, Query
+from fastapi import APIRouter, Depends, HTTPException, Security
 from starlette import status
 
 from src.dependencies.auth import Scope, get_guild_client
@@ -7,7 +7,7 @@ from src.dependencies.database import db
 
 from src.models.auth import TokenPayload
 from src.models.worlds import world
-from src.models.server import items
+from src.models import items
 
 from src.services.world import WorldService
